requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
click-config-file==0.6.0
click-plugins==1.1.1.2
configobj==5.0.9
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
fonttools==4.59.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.12.3
idna==3.10
jiter==0.10.0
joblib==1.5.1
kiwisolver==1.4.8
matplotlib==3.10.5
nltk==3.9.1
numpy==2.3.2
oauthlib==3.3.1
openai==1.99.9
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
twarc==2.14.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
